DEV-29 Integrating buttons and setting up routes
