+++ src/App.tsx
@@ -5,6 +5,7 @@ import CoursePage from './pages/CoursePage/CoursePage'
 import ExampleModal from './components/Example/ExampleModal'; // Временный файл
 import MainPage from './pages/MainPage/MainPage'
 import ProfilPage from './pages/ProfilePage/ProfilePage'
+import { Route, Routes } from 'react-router-dom';
 
 function App() {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -26,14 +27,14 @@ function App() {
   return (
     <div className='bg-background min-h-screen flex flex-col px-[16px] mobil:px-[16px] desktop:px-[70px] mediumDesktop:px-[140px] pb-page-padding overflow-x-hidden'>
       <Header onLoginClick={handleLoginClick} />
-      
-      <ProfilPage />
-      
-      <MainPage />
-      
-      <CoursePage />
-      
-      <ExampleModal />  {/* Это временное решение для просмотра примера. Эту строчку, а также модуль Example, потом можно будет удалить (или не удалять, в проекте он мешать не будет)  */}
+
+      <Routes>
+        <Route path='/' element={<MainPage />}/>
+        <Route path='/user' element={<ProfilPage />}/>
+        <Route path='/course' element={<CoursePage />}/>
+
+        <Route path='/example_modal' element={<ExampleModal />}/> {/* Это временное решение для просмотра примера. Эту строчку, а также модуль Example, потом можно будет удалить (или не удалять, в проекте он мешать не будет)  */}
+      </Routes>
       
       <Modal
         isOpen={isModalOpen}
