+++ src/components/Button.tsx
@@ -17,7 +17,7 @@ const Button: React.FC<ButtonProps> = ({
   className = '',
   disabled = false,
 }) => {
-  const baseStyles = "font-roboto h-[52px] font-normal text-lg leading-[110%] text-center rounded-[46px] px-[26px] py-4 transition-all duration-300 ease-in-out cursor-pointer";
+  const baseStyles = "font-roboto h-[52px] font-normal text-lg leading-[110%] text-center rounded-[46px] transition-all duration-300 ease-in-out cursor-pointer";
 
   const variantStyles = {
     primary: "bg-primary text-black hover:bg-primary-hover active:bg-black active:text-white",
